:truck: Rename IRenderNode's file
